Convert "Library" components to typescript
