+++ src/components/library/BookmarkButton/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { FiSave as SaveIcon } from 'react-icons/fi';
-import { IconButton } from '../ui';
+import { IconButton } from 'snake-ui';
 
 const BookmarkButton = () => <IconButton icon={<SaveIcon />} color="primary" />;
